### 2022-01-23 #4094 @whedon
[PRE REVIEW]: UncertainSCI: A Python Package for Noninvasive Parametric Uncertainty Quantification of Simulation Pipelines
**Submitting author:** @jessdtate (<a href="http://orcid.org/0000-0002-2934-1453">Jess Tate</a>)
**Repository:** https://github.com/SCIInstitute/UncertainSCI
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0b
