### 2022-02-10 f734ab0
Weight assignment fix
