+++ UncertainSCI/pce.py
@@ -144,6 +144,28 @@ class PolynomialChaosExpansion():
 
         self.samples = samples
 
+    def set_weights(self):
+        """Sets weights based on assigned samples.
+        """
+        if self.samples is None:
+            raise RuntimeError("PCE weights cannot be set unless samples are set first.""")
+        
+        if self.sampling.lower() == 'greedy-induced':
+            self.weights = self.christoffel_weights()
+        elif self.sampling.lower() == 'gq':
+            M = self.sampling_options.get('M')
+            if M is None:
+                raise ValueError("The sampling option 'M' must be specified for Gauss quadrature sampling.")
+
+            _, self.weights = self.distribution.polys.tensor_gauss_quadrature(M)
+
+        elif self.sampling.lower() == 'gq-induced':
+            self.weights = self.christoffel_weights()
+
+        else:
+            raise ValueError("Unsupported sample type '{0}' for input\
+                              sample_type".format(self.sampling))
+
     def map_to_standard_space(self, q):
         """Maps parameter values from model space to standard space.
 
@@ -211,7 +233,7 @@ class PolynomialChaosExpansion():
 
                 self.samples = self.map_to_model_space(x)
 
-            self.weights = self.christoffel_weights()
+            #self.weights = self.christoffel_weights()
 
         elif self.sampling.lower() == 'gq':
 
@@ -221,7 +243,7 @@ class PolynomialChaosExpansion():
 
             p_standard, w = self.distribution.polys.tensor_gauss_quadrature(M)
             self.samples = self.map_to_model_space(p_standard)
-            self.weights = w
+            #self.weights = w
 
         elif self.sampling.lower() == 'gq-induced':
 
@@ -232,11 +254,13 @@ class PolynomialChaosExpansion():
             p_standard = self.distribution.opolys.idist_gq_sampling(K, self.indices, M=self.sampling_options.get('M'))
 
             self.samples = self.map_to_model_space(p_standard)
-            self.weights = self.christoffel_weights()
+            #self.weights = self.christoffel_weights()
 
         else:
             raise ValueError("Unsupported sample type '{0}' for input\
                               sample_type".format(self.sampling))
+
+        self.set_weights()
 
     def integration_weights(self):
         """
