### 2022-02-17 b903469
force np array for constistancy
