### 2022-02-17 05be2e1
Updated normal dist demo
