+++ demos/build_pce_normal.py
@@ -5,14 +5,15 @@ from matplotlib import pyplot as plt
 
 from UncertainSCI.distributions import NormalDistribution
 from UncertainSCI.model_examples import sine_modulation
-from UncertainSCI.indexing import TotalDegreeSet
 from UncertainSCI.pce import PolynomialChaosExpansion
-# # Distribution setup
 
+from UncertainSCI.vis import piechart_sensitivity, quantile_plot, mean_stdev_plot
+
+## Distribution setup: generating a correlated multivariate normal distribution
 # Number of parameters
 dimension = 2
 
-# Specifies normal distribution
+# Specifies statistics of normal distribution
 mean = np.ones(dimension)
 cov = np.random.randn(dimension, dimension)
 cov = np.matmul(cov.T, cov)/(4*dimension)  # Some random covariance matrix
@@ -20,178 +21,39 @@ cov = np.matmul(cov.T, cov)/(4*dimension)  # Some random covariance matrix
 # Normalize so that parameters with larger index have smaller importance
 D = np.diag(1/np.sqrt(np.diag(cov)) * 1/(np.arange(1, dimension+1)**2))
 cov = D @ cov @ D
-
 dist = NormalDistribution(mean=mean, cov=cov, dim=dimension)
 
-# # Indices setup
-order = 10
-index_set = TotalDegreeSet(dim=dimension, order=order)
-
-print('This will query the model {0:d} times'.format(index_set.get_indices().shape[0] + 10))
-# Why +10? That's the default for PolynomialChaosExpansion.build_pce_wafp
-
-# # Initializes a pce object
-pce = PolynomialChaosExpansion(index_set, dist)
-
-# # Define model
+## Define forward model
 N = int(1e2)  # Number of degrees of freedom of model
 left = -1.
 right = 1.
 x = np.linspace(left, right, N)
 model = sine_modulation(N=N)
 
-# # Three equivalent ways to run the PCE model:
+## Polynomial order
+order = 5
 
-# 1
-# Generate samples and query model in one call:
-pce = PolynomialChaosExpansion(index_set, dist)
-lsq_residuals = pce.build(model, oversampling=10)
+## Initializes a pce object
+pce = PolynomialChaosExpansion(distribution=dist, order=order, plabels=['p1', 'p2'])
+pce.generate_samples()
 
+print('This queries the model {0:d} times'.format(pce.samples.shape[0]))
+
+# Query model:
+pce.build(model)
 
 # The parameter samples and model evaluations are accessible:
 parameter_samples = pce.samples
 model_evaluations = pce.model_output
 
-# And you could build a second PCE on the same parameter samples
-pce2 = PolynomialChaosExpansion(index_set, dist)
+# And if desired you could build a second PCE on the same parameter samples
+pce2 = PolynomialChaosExpansion(distribution=dist, order=order)
 pce2.set_samples(parameter_samples)
 pce2.build(model)
 
-# pce and pce2 have the same coefficients:
-# np.linalg.norm( pce.coefficients - pce2.coefficients )
-
-# # Postprocess PCE: mean, stdev, sensitivities, quantiles
-mean = pce.mean()
-stdev = pce.stdev()
-
-
-# Power set of [0, 1, ..., dimension-1]
-variable_interactions = list(chain.from_iterable(combinations(range(dimension), r) for r in range(1, dimension+1)))
-
[…]
