### 2022-02-23 #4094 @NMontanaBrown
> Hi @NMontanaBrown @dsurujon 👋
>
> Would you be interested in and available to review this JOSS submission?

Hi, thanks for the invite. This is slightly out of my scope so my review likely be the most useful. Apologies for the slow reply!
