### 2022-05-03 #4249 @RichardClayton
Hi Kelly,

Thanks -- I am just recovering from an unpleasant bout of covid, which took
me out for 2 weeks. I am just catching up with emails and all of the other
tasks that have been piling up.

I did have a question about feedback to the authors. Is there a mechanism
to provide comments on the manuscript beyond the tickboxes on the checklist?

Best wishes,

Richard
--

On Mon, 2 May 2022 at 17:58, Kelly L. Rowland ***@***.***>
wrote:

> Hi @shahmoradi <https://github.com/shahmoradi> @RichardClayton
> <https://github.com/RichardClayton> thanks for getting started on your
> reviews - just checking in; please let me know if you have any questions or
> concerns.
>
