### 2022-05-03 #4249 @kellyrowland
Hi @RichardClayton - I'm really sorry to hear that, and I hope your recovery is progressing along as well as it can.

For more lengthy suggestions or requests, we suggest that reviewers open issues in the code's Github repository. So, you'll want to navigate to the issues page at https://github.com/SCIInstitute/UncertainSCI/issues and create new issues there for suggestions or requests. It can be helpful to link to those issues here in this review issue as well, to keep track of things.
