### 2022-05-25 #4249 @RichardClayton
Apologies that I have taken so long to complete my initial review.

I am content to check off most of the boxes for my review. However, there are some minor revisions that I would suggest. These are:

**Functionality documentation** - The documentation at https://uncertainsci.readthedocs.io/en/latest/api_docs/index.html is of good quality, and there are some nice demos and tutorials, which enable the user to understand how the API works. However I would like to have seen a more comprehensive documentation of the different functions, so that users can quickly find what they are looking for without trawling through the source code.

**State of the field** - The paper is well written, and commendably short. However, I would like to have seen some reference to other methods for UQ aside from polynomial chaos expansions.

Either in the paper or in the documentation, it would also be very helpful to include a worked example or demo of a non-intrusive UQ exercise with an existing simulator that is not built in to UncertainSCI -- maybe by expanding Figure 1 in the paper. My guess is that most users will wish to deploy UncertainSCI on this type of code, and so this type of example would be expected to improve the use experience.

Best wishes,

Richard
--
