### 2022-05-25 #4249 @kellyrowland
@RichardClayton thank you for the thorough review! I understand that just about everyone is stretched quite thin these days, and the JOSS team really appreciates the effort of reviewers.

@shahmoradi checking in again on your review status - please let me know about an update when you can. If you need to set down the review for any reason, please just let me know that so I can find another reviewer.
