### 2022-06-04 #4249 @shahmoradi
Apologies for my delayed response. The mentions of this thread has apparently been mixed and overlooked among many other daily github emails.

+ Overall the package looks good, presents substantial work that would merit publication in JOSS. However, there are still a few major issues that need authors' attention, some of which have been also raised by @RichardClayton:
- The documentation is very and looks beautiful, but there are parts that are still left as "TODO", or parts that are outdated, for example, the folder `demo` is mentioned a few times in the documentation, while apparently the corresponding folder in GitHub was later renamed to `demos`. Consequently, some of the commands in the documentation are non-functional. Some of the examples also appear to be non-functional. Could these examples be converted to Jupyter notebooks ready to run (instead of appearing in the documentation of the library)?
- The paper is well-written. There is a typo in section "Statement of the need" @ "...UncertainSCI has been use to quantify..." : use --> used
- A discussion of the "state of the field" and existing similar work appears to be missing in the paper. If there is no comparable work, the authors could simply and proudly state that UncertainCI is the first of its kind ever to exist.
- There is a documentation section named "Developer documentation", which is essential to have, but it is mainly or only focused on guidelines for tutorials. JOSS requires guidelines for all categories of interested people including those who want to: 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.
- There appears to be several tests in the GitHub respoitory. But I could not find out how I could run the tests. More importantly, there is no coverage report informing the users about the tested percentage of the codebase. This is crucial to bring confidence and reliability with using this software. Ideally the tests and coverage reports should be automated with test results continuously daily reported on the project's front page on GitHub. There are several free code coverage services available that the authors could use, that particularly provide excellent service to the Python community.

I have left the three relevant items in the checklist unmarked for now until I receive the authors' response.

Sincerely
