### 2022-06-17 ebea612
some contribution guidelines and otherwise doc improvement
