+++ docs/user_docs/support.md
@@ -2,14 +2,12 @@
 
 ## Questions?
 
-If you have questions, please ask them on our discussion board:
+If you have questions, please ask them on our public discussion board:
 <https://github.com/SCIInstitute/UncertainSCI/discussions>
 
-Or you can email our mailing list:
-TODO
+Or, you can submit questions to our private support email at <cibc-contact@sci.utah.edu>
+
 
 ## Bugs and Requests
 
-Create an issue on Github.
-
-Github: <https://github.com/SCIInstitute/UncertainSCI/issues>
+If you find any bugs or have any feature requests, please create an issue on GitHub: <https://github.com/SCIInstitute/UncertainSCI/issues>
