### 2022-08-30 #4249 @kellyrowland
Thanks for the update - the suggested mark is around 1000 words, but that's flexible. Adding more material to round out the paper should be fine, but I'd be a little wary of, say, anything approaching 1500 words.
