### 2022-10-11 4f61287
tweaks to demos
