### 2022-11-10 #4249 @kellyrowland
Hi @jessdtate -checking in here to ask if you think the paper is ready for another look by the reviewers? I see some recent commits in the `joss_feedback` branch and just wanted to touch base.
