+++ docs/user_docs/support.md
@@ -5,7 +5,18 @@
 If you have questions, please ask them on our public discussion board:
 <https://github.com/SCIInstitute/UncertainSCI/discussions>
 
-Or, you can submit questions to our private support email at <cibc-contact@sci.utah.edu>
+Or, you can ask them on our mailing list <uncertainsci@sci.utah.edu>.  To subscribe:
+ - send an email to <sympa@sci.utah.edu> with  `subscribe uncertainsci` in the body of the email.  
+ - You will receive an email requesting confirmation of the subscription.
+ - Reply (no text needed) to the email to confirm submission.  
+
+Now you're on the list and will receive updates and questions.  You can also submit questions to the list once subscribed
+
+To unsubscribe from the mailing list:
+- send an email to <sympa@sci.utah.edu> with  `unsubscribe uncertainsci` in the body of the email.  
+- You will receive a confirmation email confirming the unsubcription.  
+
+You can also submit questions to our private support email at <cibc-contact@sci.utah.edu>
 
 
 ## Bugs and Requests
