+++ README.md
@@ -11,8 +11,7 @@ A Python-based toolkit that harnesses modern techniques to estimate model and pa
 ![mac-build](https://github.com/SCIInstitute/UncertainSCI/workflows/mac-build/badge.svg)
 ![linux-build](https://github.com/SCIInstitute/UncertainSCI/workflows/linux-build/badge.svg)
 ![windows-build](https://github.com/SCIInstitute/UncertainSCI/workflows/windows-build/badge.svg)
-[![Coverage Status](https://coveralls.io/repos/SCIInstitute/UncertainSCI/badge.png)](https://coveralls.io/r/SCIInstitute/UncertainSCI)
-
+[[![Coverage Status](https://coveralls.io/repos/github/SCIInstitute/UncertainSCI/badge.svg?branch=joss_feedback)](https://coveralls.io/github/SCIInstitute/UncertainSCI?branch=joss_feedback)
 [![status](https://joss.theoj.org/papers/660d2fe53fbf67dd2714e9546251bd33/status.svg)](https://joss.theoj.org/papers/660d2fe53fbf67dd2714e9546251bd33)
 
 
