+++ .github/workflows/pythonapp.yml
@@ -38,7 +38,14 @@ jobs:
       run: |
         pip install --editable ./
         pytest --cov=UncertainSCI
+    - name: Upload coverage data to coveralls.io
+        run: coveralls --service=github
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          COVERALLS_FLAG_NAME: ${{ matrix.os }}
+          COVERALLS_PARALLEL: true
     - name: Coveralls
-      uses: coverallsapp/github-action@v2.0.0
+        uses: coverallsapp/github-action@v2.0.0
+        
       
 
