+++ requirements.txt
@@ -1,3 +1,3 @@
 matplotlib>=3.1.3
-numpy>=1.19
+numpy>=1.22
 scipy>=1.4.1
