+++ README.md
@@ -6,12 +6,12 @@ A Python-based toolkit that harnesses modern techniques to estimate model and pa
 ![UncertainSCI](docs/_static/UncertainSCI.png "UncertainSCI")
 
 
-![Linux build](https://github.com/SCIInstitute/UncertainSCI/workflows/Linux%20build/badge.svg)
+![build](https://github.com/SCIInstitute/UncertainSCI/workflows/pythonapp/badge.svg)
 
 ![mac-build](https://github.com/SCIInstitute/UncertainSCI/workflows/mac-build/badge.svg)
 ![linux-build](https://github.com/SCIInstitute/UncertainSCI/workflows/linux-build/badge.svg)
 ![windows-build](https://github.com/SCIInstitute/UncertainSCI/workflows/windows-build/badge.svg)
-[[![Coverage Status](https://coveralls.io/repos/github/SCIInstitute/UncertainSCI/badge.svg?branch=joss_feedback)](https://coveralls.io/github/SCIInstitute/UncertainSCI?branch=joss_feedback)
+[![Coverage Status](https://coveralls.io/repos/github/SCIInstitute/UncertainSCI/badge.svg?branch=master)](https://coveralls.io/github/SCIInstitute/UncertainSCI?branch=master)
 [![status](https://joss.theoj.org/papers/660d2fe53fbf67dd2714e9546251bd33/status.svg)](https://joss.theoj.org/papers/660d2fe53fbf67dd2714e9546251bd33)
 
 
