+++ README.md
@@ -6,11 +6,7 @@ A Python-based toolkit that harnesses modern techniques to estimate model and pa
 ![UncertainSCI](docs/_static/UncertainSCI.png "UncertainSCI")
 
 
-![build](https://github.com/SCIInstitute/UncertainSCI/workflows/Build/badge.svg?branch=joss_feedback)
-
-![mac-build](https://github.com/SCIInstitute/UncertainSCI/workflows/mac-build/badge.svg)
-![linux-build](https://github.com/SCIInstitute/UncertainSCI/workflows/linux-build/badge.svg)
-![windows-build](https://github.com/SCIInstitute/UncertainSCI/workflows/windows-build/badge.svg)
+![All Builds](https://github.com/SCIInstitute/UncertainSCI/workflows/Build/badge.svg)
 [![Coverage Status](https://coveralls.io/repos/github/SCIInstitute/UncertainSCI/badge.svg?branch=master)](https://coveralls.io/github/SCIInstitute/UncertainSCI?branch=master)
 [![status](https://joss.theoj.org/papers/660d2fe53fbf67dd2714e9546251bd33/status.svg)](https://joss.theoj.org/papers/660d2fe53fbf67dd2714e9546251bd33)
 
