### 2023-03-29 8be8297
paper additions per commments.  updated examples to unclude scirun.  fixed typos
