+++ paper/paper.bib
@@ -106,6 +106,22 @@ year = {2020}
   volume =       "48"
 }
 
+@Article{JDT:Rah2016,
+  author =       "A. {Rahimi} and J. {Sapp} and J. {Xu} and P.
+                 {Bajorski} and M. Hor\'{a}\v{c}ek and L. {Wang}",
+  journal =      "{IEEE} Trans. Med. Imag.",
+  title =        "Examining the Impact of Prior Models in Transmural
+                 Electrophysiological Imaging: {A} Hierarchical
+                 Multiple-Model Bayesian Approach",
+  year =         "2016",
+  volume =       "35",
+  number =       "1",
+  pages =        "229--243",
+   doi =          "10.1109/TMI.2015.2464315",
+  ISSN =         "0278-0062",
+  month =        jan,
+}
+
 
 @article{JDT:Tat2021c,
 	author = {Tate, J.D. and Rampersad, S. and Charlebois, C. and Liu, Zexin and Bergquist, J.A. and White, D. and Rupp, L.C. and Brooks, D.H. and Narayan, A. and MacLeod, R.S.},
@@ -124,6 +140,19 @@ year = {2020}
 	volume = {14},
 	year = {2021}
 	}
+	
+@Article{JDT:Xu2014,
+  author =       "Jingjia Xu and John L Sapp and Azar Rahimi Dehaghani
+                 and Fei Gao and Linwei Wang",
+  date =         "2014",
+  journal =      "Med Image Comput Comput Assist Interv",
+  number =       "Pt 2",
+  pages =        "529--537",
+  title =        "Variational Bayesian electrophysiological imaging of
+                 myocardial infarction.",
+  volume =       "17",
+  year =         "2014",
+}	
 
 @book{ACN:Xiu2010,
 	title = {Numerical {Methods} for {Stochastic} {Computations}: {A} {Spectral} {Method} {Approach}},
@@ -133,6 +162,18 @@ year = {2020}
 	author = {Xiu, D.},
 	month = jul,
 	year = {2010}
+}
+
+@incollection{ACN:Ras2004,
+  title =        "Gaussian {Processes} in {Machine} {Learning}",
+  series =       "Lecture {Notes} in {Computer} {Science}",
+  isbn =         "978-3-540-23122-6 978-3-540-28650-9",
+  booktitle =    "Advanced {Lectures} on {Machine} {Learning}",
+  publisher =    "Springer, Berlin, Heidelberg",
+  author =       "Rasmussen, C. E.",
+  year =         "2004",
+  doi =          "10.1007/978-3-540-28650-9_4",
+  pages =        "63--71",
 }
 
 @article{ACN:Guo2018,
@@ -173,3 +214,5 @@ year = {2020}
 	note = {arXiv:1704.08465 [math]},
 	pages = {71--97}
 }
+
+
