+++ paper/paper.md
@@ -55,7 +55,8 @@ bibliography: paper.bib
 
 # Summary
 
-We have developed UncertainSCI [@USCI] as an open-source tool designed to make modern uncertainty quantification (UQ) techniques more accessible in biomedical simulation applications. UncertainSCI is implemented in Python with a noninvasive interface to meet our software design goals of 1) numerical accuracy, 2) simple application programming interface (API), 3) adaptability to many applications and methods, and 4) interfacing with diverse simulation software.  Using a Python implementation in UncertainSCI allowed us to utilize the popularity and low barrier-to-entry of Python and its common packages and to leverage the built-in integration and support for Python in common simulation software packages and languages. Additionally, we used noninvasive UQ techniques and created a similarly noninvasive interface to external modeling software that can be called in diverse ways, depending on the complexity and level of Python integration in the external simulation pipeline. We have developed and included examples applying UncertainSCI to relatively simple 1D simulations implemented in Python, and to bioelectric field simulations implemented in external software packages, which demonstrate the use of UncertainSCI and the effectiveness of the architecture and implementation in achieving our design goals. Figure \autoref{fig:pipeline} illustrates the use of UncertainSCI in computing UQ with modeling pipelines for bioelectricity simulations.  
+We have developed UncertainSCI [@USCI] as an open-source tool designed to make modern uncertainty quantification (UQ) techniques more accessible in biomedical simulation applications. UncertainSCI is implemented in Python with a noninvasive interface to meet our software design goals of 1) numerical accuracy, 2) simple application programming interface (API), 3) adaptability to many applications and methods, and 4) interfacing with diverse simulation software.  Using a Python implementation in UncertainSCI allowed us to utilize the popularity and low barrier-to-entry of Python and its common packages and to leverage the built-in integration and support for Python in common simulation software packages and languages. Additionally, we used noninvasive UQ techniques and created a similarly noninvasive interface to external modeling software that can be called in diverse ways, depending on the complexity and level of Python integration in the external simulation pipeline. We have developed and included examples applying UncertainSCI to relatively simple 1D simulations implemented in Python, and to bioelectric field simulations implemented in external software packages, which demonstrate the use of UncertainSCI and the effectiveness of the architecture and implementation in achieving our design goals.  UnceratainSCI differs from similar software, notably [UQLab](https://www.uqlab.com), [Uncertainpy](https://uncertainpy.readthedocs.io/en/latest/), and [Simnibs](https://simnibs.github.io/simnibs/build/html/tutorial/advanced/uq.html), in that it can be efficiently and non-invasively used with external simulation software, specifically with high resolution 3D simulations often used in Bioelectric field simulations.  Figure \autoref{fig:pipeline} illustrates the use of UncertainSCI in computing UQ with modeling pipelines for bioelectricity simulations.  
+
 
 ![User pipeline for UncertainSCI.  After the user inputs parameter distributions, UncertainSCI will compute an efficient sampling scheme.  The parameter samples are run through the targeted modeling pipeline, which can be implemented in external software tools.  The computed solutions are collected and compiled into relevant statistics with UncertainSCI. \label{fig:pipeline}](UncertainSCI_pipeline.png){ width=100% }
 
@@ -67,7 +68,7 @@ Biomedical computer models include many input parameters that do not have precis
 
 # Mathematics
 
[…]
