+++ .github/workflows/pypublish.yml
@@ -6,9 +6,9 @@ jobs:
     name: Publish Python distribution to PyPI and TestPyPI
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - name: Setup Python 3.8
-        uses: actions/setup-python@v2
+        uses: actions/setup-python@v3
         with:
           python-version: 3.8
       - name: Install pypa/build
