+++ paper/paper.md
@@ -49,7 +49,7 @@ affiliations:
    index: 5
  - name: Electrical and Computer Engineering Department, Northeastern University, Boston, MA, USA
    index: 6
-date: 4 Jan 2022
+date: Mar 29, 2023
 bibliography: paper.bib
 ---
 
