### 2023-03-29 19a2e34
bump Usci version (1.0-rc1).  bump python versions in setup
