### 2023-04-05 52de2ab
Joss feedback (#106)
* gitignore update

* some contribution guidelines and otherwise doc improvement

* tweaks to demos

* revert demo update

* update python dependancies to avoid vulnerabilities

* update python versions

* update version number

* some edits

* contribute guide done

* a couple additions

* some typos

* mailing list, some badgers

* tryout some badgers

* coveralls action

* coveralls action fixes

* coveralls action fixes

* coveralls action fixes ii

* update python versions

* upgrade numpy for security

* upgrade numpy for security

* coverall completion check

* some
[…]
