+++ requirements.txt
@@ -1,6 +1,3 @@
-matplotlib==3.1.3; python_version < '3.8'
-matplotlib>=3.1.3; python_version >= '3.8'
-numpy==1.15.2; python_version < '3.8'
-numpy>=1.21.0; python_version >= '3.8'
-scipy==1.4.1; python_version < '3.8'
-scipy>=1.4.1; python_version >= '3.8'
+matplotlib>=3.1.3
+numpy>=1.22
+scipy>=1.4.1
