+++ docs/dev_docs/contribute.md
@@ -1,9 +1,107 @@
 # Contribution Guide
 
-Thank you for you cobributions to UncertainSCI!  
+Thank you for you contributions to UncertainSCI!  We welcome and appreciate and contributions, from reporting a bugs to code contributions.  If you wish to contribute, please do so in the following ways.
 
-If you'd like to make a tutorial, please follow the instructions in the [Tutorial Tutorial](HowToTutorial.html).  
+## Community Support 
 
-We will be filling in more directions about how to to contribute.  Try to do the best that you can.  
+A great way to start contributing to UncertainSCI is to submit and answer questions on our [discussion board]<https://github.com/SCIInstitute/UncertainSCI/discussions>.   
+
+Other ways of contacting the communtity are located on our [support page](../user_docs/support.html#support)
+
+## Bugs and Features
+
+We  encourage users to report any bugs they find and request any features they'd like as a [GitHub issue]<https://github.com/SCIInstitute/UncertainSCI/issues>.  If you would like to tackle any issues, please volunteer by commenting in the issue or [assigning yourself]<https://docs.github.com/en/issues/tracking-your-work-with-issues/assigning-issues-and-pull-requests-to-other-github-users>.  
+
+
+## Make a Tutorial
+
+If you have a tutorial you'd like to share, we'd love to have it.  We have a [Tutorial Tutorial](HowToTutorial.html) to explain how to make and contribute tutorials.
+
+
+## Contribute Code
+
+We appreciate to code maintenance and development that our community can provide.  If you'd like to submit a bug fix, dependency update, or an added feature, please keep in mind the [style guide](#style-guide), create a [fork](#fork-repo) of the UncertainSCI repo, and use a [Pull Request](#pull-requests) to add it to UncertainSCI.  
+
+It is best practice to make sure that there is  [GitHub issue]<https://github.com/SCIInstitute/UncertainSCI/issues> to describe the required changes to the code, and having these issues documented become more important with the scope of the additions and changes.  Possible additions can also be discussed on our [discussion board]<https://github.com/SCIInstitute/UncertainSCI/discussions>.  
+
+### Fork Repo
+
+With your own github account, go to the [UncertainSCI Github page](https://github.com/SCIInstitute/UncertainSCI). Click the fork button on the upper right side of the page. It will ask you where to move the fork to, chose your own account. Once the repository is forked, clone it to your local machine with the following command.  
+
+```
+$git clone https://github.com/[yourgithubaccount]/UncertainSCI.git
+```
+
+After the the code is cloned, navigate to the repository directory and add the upstream path to the original UncertainSCI repository.  
+
+```
+$git remote add upstream https://github.com/SCIInstitute/UncertainSCI.git
+```
+
+You should be able to see both your and the original repository when you use the command:     
+
+```
+$git remote -v
+```
+
+The fork is good to go, but you will need to sync the fork occasionally to keep up with the changes in the main repository.  To sync your fork, use the following commands:
+
+```
+$git fetch upstream
+
+$git checkout master
+
+$git merge upstream/master
+```
+You should sync and merge your fork before you start a new module and before you create a pull request.  
+It is a good practice to create a new branch in your fork for every module you will be adding.  The command to create a new branch is:
+
+```
+$git checkout -b [branch_name]
+```
+
+Please see the [Github help page](https://help.github.com) for more information.
+
+### Pull Requests
+
[…]

[…]
