+++ docs/.readthedocs.yaml
@@ -16,11 +16,10 @@ sphinx:
    configuration: docs/conf.py
 
 # If using Sphinx, optionally build your docs in additional formats such as PDF
- formats:
-    - pdf
-    - EPUB
+ formats: all
 
 # Optionally declare the Python requirements required to build your docs
 python:
    install:
    - requirements: docs/requirements.txt
+   - path: ./
