+++ .readthedocs.yaml
@@ -5,6 +5,7 @@
 # Required
 version: 2
 
+
 # Set the version of Python and other tools you might need
 build:
   os: ubuntu-22.04
