+++ .readthedocs.yaml
@@ -5,7 +5,6 @@
 # Required
 version: 2
 
-
 # Set the version of Python and other tools you might need
 build:
   os: ubuntu-22.04
@@ -19,7 +18,7 @@ sphinx:
   fail_on_warning: true
 
 # If using Sphinx, optionally build your docs in additional formats such as PDF
- formats: all
+formats: all
 
 # Optionally declare the Python requirements required to build your docs
 python:
