+++ .readthedocs.yaml
@@ -27,5 +27,3 @@ python:
       path: .
       extra_requirements:
         - docs
-    - method: pip
-      path: package
