### 2023-04-10 b69a0ea
try to install docs separately
