+++ .readthedocs.yaml
@@ -27,3 +27,5 @@ python:
       path: .
       extra_requirements:
         - docs
+    - method: pip
+      path: docs
