+++ docs/requirements.txt
@@ -6,5 +6,5 @@ sphinx-issues
 sphinx-markdown-tables
 sphinx-notfound-page
 sphinx_rtd_theme
-sphinxcontrib-bibtex
+sphinxcontrib
 sphinx-markdown-tables
