+++ .readthedocs.yaml
@@ -27,3 +27,6 @@ python:
       path: .
       extra_requirements:
         - docs
+#    - requirements: requirements.txt
+#    - requirements: docs/requirements.txt
+#    - path: .
