+++ docs/requirements.txt
@@ -6,5 +6,5 @@ sphinx-issues
 sphinx-markdown-tables
 sphinx-notfound-page
 sphinx_rtd_theme
-sphinxcontrib
+sphinxcontrib-bibtex
 sphinx-markdown-tables
