+++ .readthedocs.yaml
@@ -23,13 +23,12 @@ formats: all
 # Optionally declare the Python requirements required to build your docs
 python:
   install:
-    - method: pip
-      path: .
-      extra_requirements:
-        - docs
-    - method: pip
-      path: .
-      
-#    - requirements: requirements.txt
-#    - requirements: docs/requirements.txt
-#    - path: .
+#    - method: pip
+#      path: .
+#      extra_requirements:
+#        - docs
+#    - method: pip
+#      path: .
+    - requirements: requirements.txt
+    - requirements: docs/requirements.txt
+    - path: .
