+++ .readthedocs.yaml
@@ -23,12 +23,6 @@ formats: all
 # Optionally declare the Python requirements required to build your docs
 python:
   install:
-#    - method: pip
-#      path: .
-#      extra_requirements:
-#        - docs
-#    - method: pip
-#      path: .
     - requirements: requirements.txt
     - requirements: docs/requirements.txt
     - path: .
