### 2023-04-10 a9fad69
some more content for quick guide and readme
