+++ docs/user_docs/getting_started.md
@@ -19,10 +19,111 @@ UncertainSCI can be built from source code using the `setup.py` script.  To call
 
 Alternatively, you can call the script with `python setup.py install`.  pip is still called to install the dependencies, but they could be installed manually instead.
 
+### Quick Guide
+
+Users can evaluate the effect of input uncertainty on forward models with UncertainSCI's implementation of [polynomial Chaos expansion (PCE)](../api_docs/pce.html).  The pipeline for the process is shown in  \autoref{fig:qg_pipeline}.
+
+![User pipeline for UncertainSCI.  After the user inputs parameter distributions, UncertainSCI will compute an efficient sampling scheme.  The parameter samples are run through the targeted modeling pipeline, which can be implemented in external software tools.  The computed solutions are collected and compiled into relevant statistics with UncertainSCI. \label{fig:qg_pipeline}](../paper/UncertainSCI_pipeline.png){ width=70% }
+
+#### Before Using UncertainSCI
+
+In order to run UncertainSCI, it must be supplied with the input parameter distributions, a model that will run the parameters in question, and a way to collect model solutions for use in PCE.  The model itself can be implemented in Python or in another software, as appropriate, as parameter sets and solutions can pass between UncertainSCI and modeling software via hard disk.  Models with high resolution solutions should could be cost prohibitive to run with PCE, so users should consider derived metrics or solutions subsets (regions of interest) for UQ.  
+
+Users must chose the input distributions for their modeling pipeline, which may significantly impact the output distributions.  The indendent application will determine the best distributions to use, yet we suggest that the users use all information available to inform their choices.  Published literature and previously generated data are, of course, more credible, but users may need to rely on their own intuition and observed trends with the targeted model.  
+
+#### PCE computation time
+
+Since UncertainSCI uses PCE to compute UQ, it is worth noting the impact of some PCE parameters on computational time.  Mostly, the time needed to compute UQ via PCE is limited by the evalutated model, especially in bioelectric field and other 3D simulation applications.  PCE attempts to reduce computational cost by limiting the number of parameter samples are needed to estimated the model output distributions, as the fewer parameter sets needed, the few times the model needs to be run.  Since PCE is estimating a polynomial to represent the output distribution, some user choices will affect the sample size
+
+*Number of parameters* modeled in the PCE will affect the number of samples needed to estimate the model uncertainty by increasing the dimensionality of the estimated polynomial. Higher demensions require more sampling points to accurately capture, so will lead to higher computation times.  However, parameters must be evaluated in the PCE together to determine the effect of their interaction in the targeted model.  
+
+*polynomial (PCE) order* affects the PCE samples needed for UQ by defining the complexity captured by PCE.  As in 1D, polynomials with higher order are able to capture higher variability within the domain.  Therefore, models with high complexity, i.e., significant response to variation in the parameter space, should use higher polynomial orders. However, more parameter samples are required to estimate higher polynomials, increasing the number of times the model must be run.  
+
+Distribution type of the parameters may effect the number of samples, but to a minor level when compared to number of parameters and polynomial order.  
+
+
+#### Running UncertainSCI for Sample Generation
+
[…]
