+++ .readthedocs.yaml
@@ -15,7 +15,7 @@ build:
 sphinx:
   builder: html
   configuration: docs/conf.py
-#  fail_on_warning: true
+  fail_on_warning: true
 
 # If using Sphinx, optionally build your docs in additional formats such as PDF
 formats: all
