+++ docs/user_docs/getting_started.md
@@ -45,7 +45,7 @@ Distribution type of the parameters may effect the number of samples, but to a m
 #### Running UncertainSCI for Sample Generation
 
 With the model setup, the user will need to setup the input parameter distribution using UncertainSCI's [distribution datatype](../api_docs/distributions.html).  While there are a few distribution types to choose from, the [Beta distribution](../api_docs/distributions.html#UncertainSCI.distributions.BetaDistribution) is a common choice.  If we had three input parameters, we can define a different beta distribution for each, thus:
-```python
+```Python
 from UncertainSCI.distributions import BetaDistribution
 
 Nparams = 3
@@ -57,12 +57,12 @@ p3 = BetaDistribution(alpha=1., beta=1.)
 For the default case used in here, the range of each parameter is `[0 , 1]`.   
 
 After the input parameters are set, we can generate a parameter sampling that will most efficiently estimate the UQ.  In order to do this, the user must create a PCE object with the distributions.  However, the user must also define a polynomial order.  This defined as an integer:
-```python
+```Python
 # # Polynomial order
 order = 5
 ```
 Now we can create the [PCE object](../api_docs/pce.html#module-UncertainSCI.pce) and generate parameter samples:
-```python
+```Python
 from UncertainSCI.pce import PolynomialChaosExpansion
 
 # Generate samples first, then manually query model, then give model output to pce.
@@ -70,7 +70,7 @@ pce = PolynomialChaosExpansion(distribution=[p1, p2, p3], order=order, plabels=p
 pce.generate_samples()
 ```
 To use the samples generated by this PCE instantiation, users can access it through the object:
-```python
+```Python
 import numpy as np # output is np array
 pce.samples
 ```
@@ -79,17 +79,17 @@ Which returns a numpy array of size `MxN` where `N` is the number of parameters 
 #### Running Parameter Samples for Model Outputs
 
 While this is heavily dependent on the modeling pipeline, the generated samples can be save to disk to run in an external pipeline:
-```python
+```Python
 np.savetxt(filename,pce.samples)
 ```
 or called within python:
-```python
+```Python
 for ind in range(pce.samples.shape[0]):
     model_output[ind, :] = model(pce.samples[ind, :])
 ```
 
 If the samples are saved to disk to run asyncronously, they will need to be added to a PCE object if the original one is destroyed.  This happens when it is run as a script, then rerun with data and is acheived with:
-```python
+```Python
 pce.set_samples(np.loadtxt(filename))
 ```
 instead of running `pce.generate_samples()`.  Similarly, the output results from the simulation will need to be a numpy array of `MxP` where `M` is the number of parameter sets, and `P` is the size of the solution array.  2D and 3D solutions can be flattened into a 1D array and collated into this solution array, then reshaped for visualization.    
@@ -97,11 +97,11 @@ instead of running `pce.generate_samples()`.  Similarly, the output results from
 #### PCE and the Output Statistics
 
 With the appropriate distributions and samples added to the PCE object and model output collect, the estimator and output statistics can be generated.  First, the PCE must be built:
-```python
+```Python
 pce.build(model_output=model_output)
 ```
 Then, output statistics can be return:
-```python
+```Python
 mean = pce.mean()
 stdev = pce.stdev()
 global_sensitivity, variable_interactions = pce.global_sensitivity()
