+++ docs/index.rst
@@ -13,10 +13,10 @@ Welcome to UncertainSCI's documentation!
 About UncertainSCI
 ===================
 
-UncertainSCI is a Python-based toolkit that harnesses modern techniques to estimate model and parametric uncertainty, with a particular emphasis on needs for biomedical simulations and applications. This toolkit enables non-intrusive integration of these techniques with well-established biomedical simulation software.
+UncertainSCI :cite:p:`JDT:Nar2022` is a Python-based toolkit that harnesses modern techniques to estimate model and parametric uncertainty, with a particular emphasis on needs for biomedical simulations and applications. This toolkit enables non-intrusive integration of these techniques with well-established biomedical simulation software.
 
 
-Currently implemented in UncertainSCI is Polynomial Chaos Expansion (PCE) with a number of input distributions.  For more information about these techniques, see: :cite:p:`JDT:Bur2020,narayan_computation_2018,guo_weighted_2018,cohen_optimal_2017`
+Currently implemented in UncertainSCI is Polynomial Chaos Expansion (PCE) with a number of input distributions.  For more information about these techniques, see: :cite:p:`JDT:Bur2020,narayan_computation_2018,guo_weighted_2018,cohen_optimal_2017`.  For studies using UncertainSCI, see: :cite:p:`JDT:DT:Ber2021,JDT:Rup2020,JDT:Rup2021,JDT:Tat2021a,JDT:Tat2021c,JDT:Tat2022`
 
 
 .. toctree::
