+++ docs/index.rst
@@ -16,7 +16,7 @@ About UncertainSCI
 UncertainSCI :cite:p:`JDT:Nar2022` is a Python-based toolkit that harnesses modern techniques to estimate model and parametric uncertainty, with a particular emphasis on needs for biomedical simulations and applications. This toolkit enables non-intrusive integration of these techniques with well-established biomedical simulation software.
 
 
-Currently implemented in UncertainSCI is Polynomial Chaos Expansion (PCE) with a number of input distributions.  For more information about these techniques, see: :cite:p:`JDT:Bur2020,narayan_computation_2018,guo_weighted_2018,cohen_optimal_2017,gupta1983`.  For studies using UncertainSCI, see: :cite:p:`JDT:DT:Ber2021,JDT:Rup2020,JDT:Rup2021,JDT:Tat2021a,JDT:Tat2021c,JDT:Tat2022`
+Currently implemented in UncertainSCI is Polynomial Chaos Expansion (PCE) with a number of input distributions.  For more information about these techniques, see: :cite:p:`JDT:Bur2020,narayan_computation_2018,guo_weighted_2018,cohen_optimal_2017,gupta1983`.  For studies using UncertainSCI, see: :cite:p:`JDT:Ber2021,JDT:Rup2020,JDT:Rup2021,JDT:Tat2021a,JDT:Tat2021c,JDT:Tat2022`
 
 
 .. toctree::
@@ -24,6 +24,7 @@ Currently implemented in UncertainSCI is Polynomial Chaos Expansion (PCE) with a
    :caption: Contents:
 
    user_docs/getting_started
+   notebooks/build_pce
    user_docs/support
    tutorials/index
    dev_docs/index
