### 2023-04-11 e25da50
move notebooks back to try out
