+++ docs/tutorials/demos/index.rst
@@ -5,6 +5,5 @@ Demos
    :maxdepth: 2
    :caption: Contents:
 
-   notebook_links
    simple_boxplots
    pce_statistics
