+++ docs/tutorials/index.rst
@@ -5,7 +5,9 @@ Tutorials
    :maxdepth: 2
    :caption: Contents:
 
+   build_pce
    demos/index
+   adapt_pce
    models
    parameters
    template
