### 2023-04-12 a17c965
flushing out ipynbs,  Getting adapt_PCE to animate (in jupyter), more links in docs, typos in pce class docs
