+++ docs/user_docs/getting_started.md
@@ -17,13 +17,13 @@ The source code for UncertainSCI can be downloaded from the [Github page](https:
 
 UncertainSCI can be built from source code using the `setup.py` script.  To call this script, navagate to the source root directory in a terminal window and run the command `pip install .` or `python -m pip install .`  
 
-### Quick Guide
+## UncertainSCI Overview
 
 Users can evaluate the effect of input uncertainty on forward models with UncertainSCI's implementation of [polynomial Chaos expansion (PCE)](../api_docs/pce.html).  The pipeline for the process is shown in the following image:
 
 ![User pipeline for UncertainSCI.  After the user inputs parameter distributions, UncertainSCI will compute an efficient sampling scheme.  The parameter samples are run through the targeted modeling pipeline, which can be implemented in external software tools.  The computed solutions are collected and compiled into relevant statistics with UncertainSCI.](../../paper/UncertainSCI_pipeline.png)
 
-#### Before Using UncertainSCI
+### Before Using UncertainSCI
 
 In order to run UncertainSCI, it must be supplied with the input parameter distributions, a model that will run the parameters in question, and a way to collect model solutions for use in PCE.  The model itself can be implemented in Python or in another software, as appropriate, as parameter sets and solutions can pass between UncertainSCI and modeling software via hard disk.  Models with high resolution solutions should could be cost prohibitive to run with PCE, so users should consider derived metrics or solutions subsets (regions of interest) for UQ.  
 
@@ -38,6 +38,9 @@ Since UncertainSCI uses PCE to compute UQ, it is worth noting the impact of some
 *polynomial (PCE) order* affects the PCE samples needed for UQ by defining the complexity captured by PCE.  As in 1D, polynomials with higher order are able to capture higher variability within the domain.  Therefore, models with high complexity, i.e., significant response to variation in the parameter space, should use higher polynomial orders. However, more parameter samples are required to estimate higher polynomials, increasing the number of times the model must be run.  
 
 Distribution type of the parameters may effect the number of samples, but to a minor level when compared to number of parameters and polynomial order.  
+
+
+### Running UncertainSCI
 
 
 #### Running UncertainSCI for Sample Generation
@@ -120,8 +123,8 @@ The [API documentation](../api_docs/index.html) explains the implementation of U
 
 ## Running UncertainSCI Demos
 
-There are a number of demos included with UncertainSCI to test it's installation and to demonstrate its use.  
+There are a number of demos included with UncertainSCI to test it's installation and to demonstrate its use.  The [previous description](#running-uncertainsci) can be found as a notebook with some more details [here](../tutorials/notebooks/build_pce.html) and as a script in `demos/build_pce.py`.  
 
-`demos/build_pce.py`  is particularly useful for a quick demonstration of parametric uncertainty quantification using polynomial chaos methods.  To run this demo, make sure that UncertainSCI is [installed](#getting-uncertainsci), then simply call the script with python using the command `python demos/build_pce.py`.  Other demos can be run similarily.  
+The demo scripts can be a way to quickly test the installation.  Make sure that UncertainSCI is [installed](#getting-uncertainsci), then simply call the script with python using the command `python demos/build_pce.py`.  Other demos can be run similarily.  
 
 We have included a number of demos and [tutorials](../tutorials/index.html) to teach users how to use UncertainSCI with various examples.  
