### 2023-04-15 faa5bab
Sphinx build (#108)
* updating release version first

* some setup changes

* adding config file for rtd

* some mods to config

* moved rtd.yaml to correct place

* had to readd config

* block endings?

* typo space

* another oopsie

* remove failonwarnngs

* try to install docs separately

* didn't work

* change requirements

* try config again

* a simpler approach

* working.  now some cleanup

* some more content for quick guide and readme

* turn fail_on_warnings

* fix for blockquotes

* too many warnings

* some obsolet info and fixing image

* refs for fig not working so removed

*
[…]
