+++ docs/references.bib
@@ -7,18 +7,127 @@
 	year = {1983},
 }
 
-@article{JDT:Bur2020,
-author = {Burk, Kyle M. and Narayan, Akil and Orr, Joseph A.},
-title = {Efficient sampling for polynomial chaos-based uncertainty quantification and sensitivity analysis using weighted approximate Fekete points},
-journal = {International Journal for Numerical Methods in Biomedical Engineering},
-volume = {36},
-number = {11},
-pages = {e3395},
-keywords = {approximate Fekete points, cardiovascular modeling, oxyhemoglobin dissociation, polynomial chaos expansion, sensitivity analysis, uncertainty quantification},
-doi = {https://doi.org/10.1002/cnm.3395},
-url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/cnm.3395},
-year = {2020}
+@Article{JDT:Nar2022,
+  title =        "Uncertain{SCI}: Uncertainty Quantification for
+                 Computational Models in Biomedicine and
+                 Bioengineering",
+  author =       "Akil Narayan and Zexin Liu and Jake Bergquist and
+                 Chantel Charlebois and Sumientra Rampersad and Lindsay
+                 Rupp and Dana Brooks and Dan White and Jess Tate and
+                 Rob S MacLeod",
+  journal =      "Available at SSRN 4049696",
+  year =         "2022",
 }
+
+@Article{JDT:Bur2020,
+  author =       "Kyle M. Burk and Akil Narayan and Joseph A. Orr",
+  title =        "Efficient sampling for polynomial chaos-based
+                 uncertainty quantification and sensitivity analysis
+                 using weighted approximate Fekete points",
+  journal =      "International Journal for Numerical Methods in
+                 Biomedical Engineering",
+  volume =       "36",
+  number =       "11",
+  pages =        "e3395",
+  keywords =     "approximate Fekete points, cardiovascular modeling,
+                 oxyhemoglobin dissociation, polynomial chaos expansion,
+                 sensitivity analysis, uncertainty quantification",
+  doi =          "https://doi.org/10.1002/cnm.3395",
+  year =         "2020",
+}
+
+@InProceedings{JDT:Ber2021,
+  author =       "Jake Bergquist and Brian Zenger and Lindsay Rupp and
+                 Akil Narayan and Jess Tate and Rob MacLeod",
+  title =        "Uncertainty Quantification in Simulations of
+                 Myocardial Ischemia",
+  booktitle =    "Computing in Cardiology",
+  month =        sep,
+  year =         "2021",
+  doi =          "",
+  volume =       "48",
+}
+
+@InProceedings{JDT:Rup2020,
+  author =       "Lindsay C Rupp and Zexin Liu and Jake A Bergquist and
+                 Sumientra Rampersad and Dan White and Jess D Tate and
+                 Dana H. Brooks and Akil Narayan and Rob S. MacLeod",
+  title =        "Using Uncertain{SCI} to Quantify Uncertainty in
+                 Cardiac Simulations",
+  booktitle =    "Computing in Cardiology",
+  month =        sep,
+  year =         "2020",
+  doi =          "",
+  volume =       "47",
+}
+
+@InProceedings{JDT:Rup2021,
+  author =       "Lindsay C Rupp and Jake A Bergquist and Brian Zenger
+                 and Karli Gillette and Akil Narayan and Jess Tate and
+                 Gernot Plank and Rob S. MacLeod",
+  title =        "The Role of Myocardial Fiber Direction in Epicardial
+                 Activation Patterns via Uncertainty Quantification",
+  booktitle =    "Computing in Cardiology",
+  month =        sep,
[…]
