+++ docs/requirements.txt
@@ -8,3 +8,5 @@ sphinx-notfound-page
 sphinx_rtd_theme
 sphinxcontrib-bibtex
 sphinx-markdown-tables
+nbsphinx
+ipykernel
