### 2023-04-15 #4249 @jessdtate
I've made a new tag for review: <https://github.com/SCIInstitute/UncertainSCI/releases/tag/v1.0rc2> .  Reviewers can use the master branch, the linked tag,  or the equivalent Pypi version to test.  Key changes include:

- correctly build API documentation so that it includes functions and descriptions <https://uncertainsci.readthedocs.io/en/latest/api_docs/index.html>
- added a sentence to the paper mentioning other methods
- added a use case combining UncertainSCI and SCIRun (<https://github.com/SCIInstitute/UQExampleBEMHeartPosition>) and linked to in the documentation (<https://uncertainsci.readthedocs.io/en/latest/tutorials/models.html#scirun-uncertainsci-ecg-uncertainty-due-to-cardiac-postion>)

- Flushed out documentation and cleaned up as many typos as posible
- fixed typos in the paper
- added a sentences in the paper mentioning other UQ software
- added a more comprehensive contribution guide.  Support page linked more prevalently
- added a testing page to documentation (<https://uncertainsci.readthedocs.io/en/latest/dev_docs/contribute.html#testing>)
- Cleaned up demos and made sure they run
- created a docs/tutorials/notebooks directory with jupyter notebook examples
