### 2023-04-16 #4249 @shahmoradi
@jessdtate
Thank you for your efforts and revision.  I attempted to run the tests last night, but finishing them took more than 30 minutes without any reports, which was longer than my patience. I ended up canceling it because the tests used up the CPU at a high clock rate for a long time.
1. How long do these tests take to complete?
1. Is there a coverage report for these tests to verify the comprehensiveness of the tests? Setting up an automated Code Coverage report generator would be an ideal solution.
