### 2023-04-16 #4249 @jessdtate
The test shouldn't take much longer than that, but we are using a coverage test service:
<https://coveralls.io/github/SCIInstitute/UncertainSCI?branch=master>.  There is a badge on the readme to show the coverage too
