### 2023-05-03 2e8e26e
adding pytests to github actions
