+++ .github/workflows/pythonapp.yml
@@ -33,11 +33,14 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-    - name: Test with pytest
+    - name: Run tests with pytest
       continue-on-error: true
       run: |
         pip install --editable ./
         pytest
+    - name: Test coverage with pytest
+      continue-on-error: true
+      run: |
         pytest --cov=UncertainSCI
     - name: Upload coverage data to coveralls.io
       run: |
