+++ .github/workflows/pythonapp.yml
@@ -33,10 +33,12 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+    - name: Test build
+      run: |
+        pip install --editable ./
     - name: Run tests with pytest
       continue-on-error: true
       run: |
-        pip install --editable ./
         pytest
     - name: Test coverage with pytest
       continue-on-error: true
