+++ .github/workflows/pythonapp.yml
@@ -36,6 +36,7 @@ jobs:
     - name: Test build
       run: |
         pip install --editable ./
+      env:
   tests:
     name: test code and coverage
     needs: build
@@ -64,8 +65,8 @@ jobs:
     steps:
     - name: Finished
       run: |
-        pip3 install --upgrade pip
-        pip3 install --upgrade coveralls
+#        pip3 install --upgrade pip
+#        pip3 install --upgrade coveralls
         coveralls --service=github --finish
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
