### 2023-05-03 #4249 @jessdtate
another note on the tests, they should only take a couple minutes to run.  I've added them to the build test on github actions under `Run tests with pytest`.   it takes about a minute on those resources, and it is setup for multiplatform.  Here is a link to the [latest run](https://github.com/SCIInstitute/UncertainSCI/actions/runs/4875480673/jobs/8697731884#step:8:78), which should be available for a few weeks.
