### 2023-06-13 #4249 @shahmoradi
My apologies again for missing this review among many GitHub notifications. Reminders are always helpful, and thank you for reminding me. I have marked all items in the checklist as complete. Test coverage of 50%, as reflected in the repository badge, is quite a dismal coverage rate for a good package. Authors should strive to bring it above 90% to ensure quality and correctness. But I trust the developers' ability and commitment to do so without delaying the JOSS publication.
