### 2023-07-05 #4249 @kellyrowland
@jessdtate this looks good to me. I will note that the various links in the "Overview" section of the README are broken, but the standalone user documentation is quite comprehensive and I don't think it's a blocker for paper acceptance.

At this point I will ask you to issue a new tagged release of the software (if changed since the start of the review), and archive it (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or elsewhere). Then, please then post the version number and archive DOI here in the review issue, and I'll follow some subsequent wrap-up steps from there.
