### 2023-08-07 #4249 @kellyrowland
hi @jessdtate pinging on this in case you're able to move forward with the wrap up steps:

>At this point I will ask you to issue a new tagged release of the software (if changed since the start of the review), and archive it (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or elsewhere). Then, please then post the version number and archive DOI here in the review issue, and I'll follow some subsequent wrap-up steps from there.
