### 2023-08-08 ee2dd0a
adding pytests to github actions (#109)
* adding pytests to github actions

* some moving around

* some more moving around

* try spiting it up a bit

* some errors

* some errors

* some other errors

* some other errors

* working now?

* undo some changes
