+++ .github/workflows/pythonapp.yml
@@ -33,10 +33,16 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-    - name: Test with pytest
-      continue-on-error: true
+    - name: Test build
       run: |
         pip install --editable ./
+    - name: Run tests with pytest
+      continue-on-error: true
+      run: |
+        pytest
+    - name: Test coverage with pytest
+      continue-on-error: true
+      run: |
         pytest --cov=UncertainSCI
     - name: Upload coverage data to coveralls.io
       run: |
@@ -46,11 +52,13 @@ jobs:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         COVERALLS_FLAG_NAME: ${{ matrix.os }}
         COVERALLS_PARALLEL: true
+        
   coveralls:
     name: confirm coveralls upload
     needs: build
     runs-on: ubuntu-latest
     container: python:3-slim
+    
     steps:
     - name: Finished
       run: |
