### 2023-08-08 c400cee
Create CITATION.cff (#111)
