+++ CITATION.cff
@@ -1,0 +1,36 @@
+cff-version: 1.0
+message: "If you use this software, please cite it as below."
+authors:
+  - family-names: Tate
+    given-names: Jess
+    orcid: https://orcid.org/0000-0002-2934-1453 
+  - family-names: Liu
+    given-names: Zexin
+    orcid: https://orcid.org/0000-0003-3409-5709
+  - family-names: Bergquist
+    given-names: Jake A
+  - name: Jake A Bergquist
+    orcid: https://orcid.org/0000-0002-4586-6911
+  - family-names: Rampersad
+    given-names: Sumientra
+    orcid: https://orcid.org/0000-0001-9860-4459
+  - family-names: White
+    given-names: Dan
+  - family-names: Charlebois
+    given-names: Chantel
+    orcid: https://orcid.org/0000-0002-4139-3539
+  - family-names: Rupp
+    given-names: Lindsay
+    orcid: https://orcid.org/0000-0002-2688-7688
+  - family-names: Brooks
+    given-names: Dana H
+    orcid: https://orcid.org/0000-0003-3231-6715
+  - family-names: MacLeod
+    given-names: Rob S
+    orcid: https://orcid.org/0000-0002-0000-0356
+  - family-names: Narayan
+    given-names: Akil
+    orcid: https://orcid.org/0000-0002-5914-4207
+title: SCIInstitute/UncertainSCI: v1.0
+version: v1.0
+date-released: 2023-08-08
