### 2023-08-08 #4249 @jessdtate
Hi @kellyrowland, Thanks for the reminder.   I've make a release ( [v1.0](https://github.com/SCIInstitute/UncertainSCI/releases/tag/v1.0) ) and DOI: 10.5281/zenodo.8226383.  Let me know what next steps that you'd like me to do.
