+++ .github/workflows/pypublish.yml
@@ -33,7 +33,7 @@ jobs:
         run: >-
           check-wheel-contents
           dist/*.whl
-      - uses: actions/upload-artifact@release/latest
+      - uses: actions/upload-artifact@latest
         with:
           path: ./dist/*
           
@@ -48,13 +48,13 @@ jobs:
       id-token: write
     if: startsWith(github.event.ref, 'refs/tags')
     steps:
-      - uses: actions/download-artifact@release/latest
+      - uses: actions/download-artifact@latest
         with:
           name: artifact
           path: dist
       - name: Upload to TestPyPI
-        uses: pypa/gh-action-pypi-publish@release/v1
+        uses: pypa/gh-action-pypi-publish@v1
         with:
           repository_url: https://test.pypi.org/legacy/
       - name: Upload to PyPI
-        uses: pypa/gh-action-pypi-publish@release/v1
+        uses: pypa/gh-action-pypi-publish@v1
