+++ .github/workflows/pypublish.yml
@@ -33,7 +33,7 @@ jobs:
         run: >-
           check-wheel-contents
           dist/*.whl
-      - uses: actions/upload-artifact@latest
+      - uses: actions/upload-artifact@v3
         with:
           path: ./dist/*
           
@@ -48,7 +48,7 @@ jobs:
       id-token: write
     if: startsWith(github.event.ref, 'refs/tags')
     steps:
-      - uses: actions/download-artifact@latest
+      - uses: actions/download-artifact@v3
         with:
           name: artifact
           path: dist
