+++ .github/workflows/pypublish.yml
@@ -53,8 +53,8 @@ jobs:
           name: artifact
           path: dist
       - name: Upload to TestPyPI
-        uses: pypa/gh-action-pypi-publish@v1
+        uses: pypa/gh-action-pypi-publish@release/v1
         with:
           repository_url: https://test.pypi.org/legacy/
       - name: Upload to PyPI
-        uses: pypa/gh-action-pypi-publish@v1
+        uses: pypa/gh-action-pypi-publish@release/v1
