### 2023-08-25 d68012d
fix bibliography example
