+++ docs/conf.py
@@ -89,9 +89,10 @@ language = 'python'
 on_rtd = os.environ.get('READTHEDOCS', None) == 'True'
 
 if not on_rtd:  # only import and set the theme if we're building docs locally
-    import sphinx_rtd_theme
-    html_theme = 'sphinx_rtd_theme'
-    html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
+  import sphinx_rtd_theme
+
+html_theme = 'sphinx_rtd_theme'
+html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
@@ -112,14 +113,14 @@ html_theme_options = {
 
 
 def setup(app):
-    app.add_config_value('recommonmark_config', {
-        #'url_resolver': lambda url: github_doc_root + url,
-        'auto_toc_tree_section': 'Contents',
-        'enable_math': False,
-        'enable_inline_math': False,
-        'enable_eval_rst': True,
-    }, True)
-    app.add_transform(AutoStructify)
+  app.add_config_value('recommonmark_config', {
+      #'url_resolver': lambda url: github_doc_root + url,
+      'auto_toc_tree_section': 'Contents',
+      'enable_math': False,
+      'enable_inline_math': False,
+      'enable_eval_rst': True,
+  }, True)
+  app.add_transform(AutoStructify)
 
 
 
