### 2023-08-30 2dfa7e4
messing around with bibtex stuff.  there seems to be a bug in sphinx-contrib.bibtex, so moved the bibliography to the end of the page until someone finds a fix
