+++ docs/dev_docs/HowToTutorial.md
@@ -210,10 +210,10 @@ The whole paragraph will need to be in the eval_rst block :cite:p:`JDT:Bur2020`.
 The whole paragraph will need to be in the eval_rst block :cite:p:`JDT:Bur2020`. For multiple references: :cite:p:`JDT:Bur2020,gupta1983`
 ```
 
-add a [bibliography section at the end of the file](#bibliography)
+add a bibliography section
 ````
 ```eval_rst
-.. bibliography::
+.. bibliography:: ../references.bib
 ```
 ````
 
@@ -320,7 +320,9 @@ With the embed link, just include the html into the markdown file.
 
 Datasets should be located serperately, unless the size is small.  Please [ask](https://github.com/SCIInstitute/UncertainSCI/discussions) if you have any questions. 
 
-## Bibliography
+
+
+### Bibliography
 ```eval_rst
-.. bibliography::
+.. bibliography:: ../references.bib
 ```
