### 2023-08-30 03996a5
bug fixed in docutils 0.20.1
