+++ docs/index.rst
@@ -40,6 +40,12 @@ Acknowledgements
 This project was supported by grants from the National Institute of Biomedical Imaging and Bioengineering (U24EB029012) from the National Institutes of Health.
 
 
+Bibliography
+================
+.. bibliography:: ./references.bib
+  :cited:
+
+
 Indices and Tables
 ==================
 
@@ -47,9 +53,4 @@ Indices and Tables
 * :ref:`modindex`
 * :ref:`search`
 
-
-Bibliography
-================
-.. bibliography:: ./references.bib
-  :cited:
 
