+++ docs/conf.py
@@ -95,8 +95,9 @@ language = 'python'
 
 on_rtd = os.environ.get('READTHEDOCS', None) == 'True'
 
-
-import sphinx_rtd_theme
+if not on_rtd:
+  import sphinx_rtd_theme
+  
 html_theme = 'sphinx_rtd_theme'
 html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
 
