### 2023-08-30 #4249 @kellyrowland
Looks like the `JDT:Xu2014` bibliography entry could use an addition of `doi = "10.1007/978-3-319-10470-6_66",` as well.
