### 2023-09-02 7bd6f60
dependency conflict with sphinx_rtd_theme and docutils.  reverting docutils and moving bibliography for work around.  TODO: update dependencies when they are more compatable
