+++ docs/requirements.txt
@@ -1,11 +1,11 @@
-docutils>=0.20.1
+docutils
 pygments
-recommonmark>= 0.7.1
-sphinx>=7.2.4
+recommonmark
+sphinx
 sphinx-issues
 sphinx-markdown-tables
 sphinx-notfound-page
-sphinx_rtd_theme
+sphinx_rtd_theme>=1.3.0
 sphinxcontrib-bibtex
 sphinx-markdown-tables
 nbsphinx
