### 2023-09-20 3ed3456
readme links fix (#110)
* readme links fix

* zenodo doi

* Create CITATION.cff (#111)

* change setup versions

* updated pypi upload to for new security

* reworking pypi publish ga"

* action typos

* action typos

* change some names

* more typos

* messing with versions

* messing with versions

* bumping version number to patch

* fix bibliography example

* found a typo

* adjust to change in rtd setup

* trying to fix bibliography, which is generating an extra <\div>

* messing around with bibtex stuff.  there seems to be a bug in sphinx-contrib.bibtex, so moved the bibliography to th
[…]
