+++ docs/references.bib
@@ -44,7 +44,6 @@
   booktitle =    "Computing in Cardiology",
   month =        sep,
   year =         "2021",
-  doi =          "",
   volume =       "48",
 }
 
@@ -57,7 +56,6 @@
   booktitle =    "Computing in Cardiology",
   month =        sep,
   year =         "2020",
-  doi =          "",
   volume =       "47",
 }
 
@@ -70,7 +68,6 @@
   booktitle =    "Computing in Cardiology",
   month =        sep,
   year =         "2021",
-  doi =          "",
   volume =       "48",
 }
 
@@ -123,7 +120,6 @@
                  Propagation Models",
   year =         "2022",
   volume =       "498",
-  number =       "",
   pages =        "1--4",
   doi =          "10.22489/CinC.2022.419",
 }
