+++ docs/requirements.txt
@@ -5,7 +5,7 @@ sphinx
 sphinx-issues
 sphinx-markdown-tables
 sphinx-notfound-page
-sphinx_rtd_theme
+sphinx_rtd_theme>=1.3.0
 sphinxcontrib-bibtex
 sphinx-markdown-tables
 nbsphinx
