+++ docs/index.rst
@@ -39,10 +39,8 @@ Acknowledgements
 
 This project was supported by grants from the National Institute of Biomedical Imaging and Bioengineering (U24EB029012) from the National Institutes of Health.
 
-
-Bibliography
-================
-.. bibliography::
+..
+  TODO: move bibliography back here when the docutils and sphinx_rtd_theme play nicely together
 
 
 Indices and Tables
@@ -51,3 +49,10 @@ Indices and Tables
 * :ref:`genindex`
 * :ref:`modindex`
 * :ref:`search`
+
+
+Bibliography
+================
+.. bibliography:: ./references.bib
+  :cited:
+
