### 2023-09-20 296a47e
split up workflows to make checks easier
