+++ .github/workflows/pypackage.yml
@@ -1,0 +1,39 @@
+name: Build and Test Package
+on: push
+
+jobs:
+  build-wheel:
+    name: build wheel for testing
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - name: Setup Python 3.9
+        uses: actions/setup-python@v3
+        with:
+          python-version: 3.9
+      - name: Install pypa/build
+        run: >-
+          python -m
+          pip install
+          build
+          --user
+      - name: Build a binary wheel and a source tarball
+        run: >-
+          python -m
+          build
+          --sdist
+          --wheel
+          --outdir dist/
+      - name: install check-wheel-contents
+        run: >-
+          python3 -m
+          pip install
+          check-wheel-contents
+      - name: check wheel
+        run: >-
+          check-wheel-contents
+          dist/*.whl
+      - uses: actions/upload-artifact@v3
+        with:
+          path: ./dist/*
+          
