### 2023-09-29 #4249 @kellyrowland
@arfon I would agree that this paper is just about ready to accept. I am mildly concerned about the discrepancies between the paper and the archive entry that I mentioned above, but if the EiCs don't see that as a blocker for acceptance then we can move forward.
