### 2023-10-02 ba6bddd
remove erroneous isbn from JDT:Tat2021c in paper.bib
