+++ paper/paper.bib
@@ -128,7 +128,6 @@ year = {2020}
 	booktitle = {Brain Stimulation: Basic, Translational, and Clinical Research in Neuromodulation},
 	date = {2021/11/01},
 	doi = {10.1016/j.brs.2021.10.226},
-	isbn = {1935-861X},
 	journal = {Brain Stimulation: Basic, Translational, and Clinical Research in Neuromodulation},
 	month = {2022/01/04},
 	number = {6},
