+++ .github/workflows/pypackage.yml
@@ -1,9 +1,9 @@
-name: Build and Test Package
-on: push
+name: Build Wheel
+on: [workflow_call]
 
 jobs:
   build-wheel:
-    name: build wheel for testing
+    name: Build Wheel for Testing
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
