### 2023-10-11 fb3b8de
homogenizing some of the workflows
