### 2023-10-11 7448873
bump version number, add missing doi
