+++ paper/paper.bib
@@ -149,6 +149,7 @@ year = {2020}
   pages =        "529--537",
   title =        "Variational Bayesian electrophysiological imaging of
                  myocardial infarction.",
+  doi = "10.1007/978-3-319-10470-6_66",             
   volume =       "17",
   year =         "2014",
 }	
