+++ CITATION.cff
@@ -1,4 +1,4 @@
-cff-version: 1.0.1rc1
+cff-version: 1.0.1rc2
 message: "If you use this software, please cite it as below."
 authors:
   - family-names: Tate
