+++ paper/paper.bib
@@ -160,6 +160,7 @@ year = {2020}
 	shorttitle = {Numerical {Methods} for {Stochastic} {Computations}},
 	publisher = {Princeton University Press},
 	author = {Xiu, D.},
+	doi = {10.1007/978-3-319-10470-6_66}, 
 	month = jul,
 	year = {2010}
 }
