### 2023-10-11 #4249 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
