### 2023-10-11 ab296d3
update cif to match zenodo.  typo in affiliations
