+++ .github/workflows/pypackage.yml
@@ -1,6 +1,14 @@
 name: Build Wheel
-on: [workflow_call]
-
+on:
+  workflow_call:
+    inputs:
+      config-path:
+        required: true
+        type: string
+    secrets:
+      envPAT:
+        required: true
+        
 jobs:
   build-wheel:
     name: Build Wheel for Testing
