+++ .github/workflows/testpackage.yml
@@ -3,5 +3,5 @@ on: [push, pull_request]
 
 jobs:
   buildwheel:
-    uses: .github/workflows/pypackage.yml
+    uses: ./.github/workflows/pypackage.yml
 
