+++ .github/workflows/pypackage.yml
@@ -1,13 +1,13 @@
 name: Build Wheel
-on:
-  workflow_call:
-    inputs:
-      config-path:
-        required: true
-        type: string
-    secrets:
-      envPAT:
-        required: true
+on: [workflow_call]
+#  workflow_call:
+#    inputs:
+#      config-path:
+#        required: true
+#        type: string
+#    secrets:
+#      envPAT:
+#        required: true
         
 jobs:
   build-wheel:
