+++ .github/workflows/pypublish.yml
@@ -21,12 +21,11 @@ jobs:
         with:
           name: artifact
           path: dist
-#          configuration-path: ${{ inputs.config-path }}
-#          secrets:
-#            envPAT: ${{ secrets.envPAT }}
+          configuration-path: ${{ inputs.config-path }}
+          secrets: inherit
       - name: Upload to TestPyPI
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
-          repository_url: https://test.pypi.org/legacy/
+          repository-url: https://test.pypi.org/legacy/
       - name: Upload to PyPI
         uses: pypa/gh-action-pypi-publish@release/v1
