+++ .github/workflows/pypackage.yml
@@ -44,6 +44,4 @@ jobs:
       - uses: actions/upload-artifact@v3
         with:
           path: ./dist/*
-#          repo-token: ${{ secrets.token }}
-#          configuration-path: ${{ inputs.config-path }}
           
