+++ .github/workflows/pythonapp.yml
@@ -1,6 +1,6 @@
 name: Build
 
-on: [push, pull_request]
+on: #[push, pull_request]
 
 jobs:
   build:
