+++ .github/workflows/pypublish.yml
@@ -17,19 +17,22 @@ jobs:
     permissions:
       id-token: write
     steps:
-      - uses: actions/download-artifact@v3
-        with:
-          name: artifact
-          path: dist
+#      - uses: actions/download-artifact@v3
+#        with:
+#          name: artifact
+#          path: dist
       - name: Upload to TestPyPI
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
+          packages-dir: dist
           repository-url: https://test.pypi.org/legacy/
           password: ${{ secrets.TESTPYPI_API_TOKEN }}
 #          configuration-path: ${{ inputs.config-path }}
 #          secrets: inherit
       - name: Upload to PyPI
         uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          packages-dir: dist
 #        with:
 #          configuration-path: ${{ inputs.config-path }}
 #          secrets: inherit
