+++ .github/workflows/pypublish.yml
@@ -24,7 +24,7 @@ jobs:
       - name: Upload to TestPyPI
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
-          packages-dir: dist/actions
+          packages-dir: dist
           repository-url: https://test.pypi.org/legacy/
           password: ${{ secrets.TESTPYPI_API_TOKEN }}
 #          configuration-path: ${{ inputs.config-path }}
@@ -32,7 +32,7 @@ jobs:
       - name: Upload to PyPI
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
-          packages-dir: dist/actions
+          packages-dir: dist
 #        with:
 #          configuration-path: ${{ inputs.config-path }}
 #          secrets: inherit
