+++ .github/workflows/pypublish.yml
@@ -7,13 +7,13 @@ on:
 jobs:
   build-wheel:
     uses: ./.github/workflows/pypackage.yml
-  upload_pypi:
-    name: Publish build to TestPyPI and PyPI
+  upload_testpypi:
+    name: Publish build to TestPyPI
     needs: [build-wheel]
     runs-on: ubuntu-latest
     environment:
       name: pypi
-      url: https://pypi.org/p/UncertainSCI
+      url: https://test.pypi.org/p/UncertainSCI
     permissions:
       id-token: write
     steps:
@@ -25,14 +25,21 @@ jobs:
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
           packages-dir: dist
-          repository-url: https://test.pypi.org/legacy/
-          password: ${{ secrets.TESTPYPI_API_TOKEN }}
-#          configuration-path: ${{ inputs.config-path }}
-#          secrets: inherit
+  upload_pypi:
+    name: Publish build to PyPI
+    needs: [build-wheel, upload_testpypi]
+    runs-on: ubuntu-latest
+    environment:
+      name: pypi
+      url: https://pypi.org/p/UncertainSCI
+    permissions:
+      id-token: write
+    steps:
+      - uses: actions/download-artifact@v3
+        with:
+          name: artifact
+          path: dist
       - name: Upload to PyPI
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
           packages-dir: dist
-#        with:
-#          configuration-path: ${{ inputs.config-path }}
-#          secrets: inherit
