+++ .github/workflows/testpackage.yml
@@ -1,5 +1,5 @@
 name: Build and Test Package
-on: #[push, pull_request]
+on: [push, pull_request]
 
 jobs:
   buildwheel:
