+++ .github/workflows/pypackage.yml
@@ -1,13 +1,5 @@
 name: Build Wheel
 on: [workflow_call]
-#  workflow_call:
-#    inputs:
-#      config-path:
-#        required: false
-#        type: string
-#    secrets:
-#      envPAT:
-#        required: false
         
 jobs:
   build-wheel:
