+++ CITATION.cff
@@ -1,4 +1,4 @@
-cff-version: 1.0.1
+cff-version: 1.1.0
 message: "If you use this software, please cite it as below."
 authors:
   - family-names: Tate
