### 2023-10-16 #4249 @oliviaguest
@jessdtate I think this is the correct DOI for the article?

> Xu, J., Sapp, J.L., Rahimi Dehaghani, A., Gao, F., Wang, L. (2014). Variational Bayesian Electrophysiological Imaging of Myocardial Infarction. In: Golland, P., Hata, N., Barillot, C., Hornegger, J., Howe, R. (eds) Medical Image Computing and Computer-Assisted Intervention – MICCAI 2014. MICCAI 2014. Lecture Notes in Computer Science, vol 8674. Springer, Cham. https://doi.org/10.1007/978-3-319-10470-6_66
