### 2023-10-17 5360c84
Deploy testing (#114)
* split up workflows to make checks easier

* bump verson for testing

* building workflow dependencies

* homogenizing some of the workflows

* bump version number, add missing doi

* another doi

* update cif to match zenodo.  typo in affiliations

* some mistakes in workflows

* working out inheritance

* try without forwarding

* trying to include token

* old api for tesst pypi

* typo

* disable other tests for now

* trying to get the right files

* try to find the artefact

* bumping version

* moveing dir

* some cleanup. separating and Upgrading test.pypi

* bum
[…]
