+++ paper/paper.bib
@@ -149,6 +149,7 @@ year = {2020}
   pages =        "529--537",
   title =        "Variational Bayesian electrophysiological imaging of
                  myocardial infarction.",
+  doi = "10.1007/978-3-319-10470-6_66",             
   volume =       "17",
   year =         "2014",
 }	
@@ -159,6 +160,7 @@ year = {2020}
 	shorttitle = {Numerical {Methods} for {Stochastic} {Computations}},
 	publisher = {Princeton University Press},
 	author = {Xiu, D.},
+	doi = {10.1007/978-3-319-10470-6_66}, 
 	month = jul,
 	year = {2010}
 }
