+++ .github/workflows/testpackage.yml
@@ -1,0 +1,6 @@
+name: Build and Test Package
+on: [push, pull_request]
+
+jobs:
+  buildwheel:
+    uses: ./.github/workflows/pypackage.yml
