### 2023-10-27 #4249 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
