### 2023-10-27 #4249 @oliviaguest
Huge thanks to the reviewers @shahmoradi, @RichardClayton, and also editor @kellyrowland! ✨  JOSS appreciates your work and effort. ✨ Also, big congratulations to the authors (@jessdtate)! 🥳 🍾
